reduced upkeep for Dark energy while being in a party by 50%
added a "Dark Tribute" which gives slightly very little bonus Dark Energy for kingdom owned towns (2 for castles, 3 for towns. doubled if the player owns the said settlements

fixed a perk of the necromancer not being working. Instead the player gives a very little amount of Dark Energy for Grave guards. (0.5 each)