bandaid solution for hireling
hireling lord turns to raisin mode if hireling is present. I think the problem here is the SettlementEncounter that is created, which set the AI back to hold mode. Manually changing this let it fall back to hold mode again.

for now just changed the serve days to 1 as a bandaid solution until we can fix this.